[test]: #23 add test_logout
[body]

[notes]

close #<issue-number>
